pyproject.toml
[project]
name = "qwen35-27b-with-opus46"
version = "0.1.0"
description = "Qwen 3.5 27B SFT fine-tuning pipeline (Unsloth + LoRA) with Opus 4.6 reasoning data"
readme = "README.md"
requires-python = ">=3.10,<3.13"
authors = [
    { name = "qwen35-27b-with-opus46 contributors" },
]
license = { text = "Apache-2.0" }

dependencies = [
    "torch==2.8.0",
    "triton>=3.3.0",
    "numpy",
    "pillow",
    "torchvision",
    "bitsandbytes",
    "xformers==0.0.32.post2",
    "transformers==5.2.0",
    "trl==0.22.2",
    "tokenizers",
    "accelerate",
    "peft",
    "datasets",
    "pandas",
    "pyarrow",
    "wandb",
    "huggingface_hub",
    "python-dotenv>=1.0.0",
    "unsloth",
    "unsloth_zoo",
]

[project.optional-dependencies]
flash = [
    "flash-linear-attention",
    "causal_conv1d==1.6.0",
]

[project.scripts]
qwen-finetune = "qwen_finetune.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qwen_finetune"]

[tool.uv]
# Some unsloth components are best installed from git (newer than PyPI).
# Keep these as explicit overrides so `uv sync` pulls the right revisions.
[tool.uv.sources]
unsloth = { git = "https://github.com/unslothai/unsloth" }
unsloth_zoo = { git = "https://github.com/unslothai/unsloth-zoo" }
